Remove datavec spark Fix json serialization Remove old tests Fix order related assertion issues

+++ deeplearning4j/deeplearning4j-zoo/src/main/java/org/deeplearning4j/zoo/model/InceptionResNetV1.java
@@ -83,7 +83,7 @@ public class InceptionResNetV1 extends ZooModel {
                         .addLayer("bottleneck", new DenseLayer.Builder().nIn(5376).nOut(embeddingSize).build(),
                                         "avgpool")
                         // Embeddings
-                        .addVertex("embeddings", new L2NormalizeVertex(new int[] {1}, 1e-10), "bottleneck")
+                        .addVertex("embeddings", new L2NormalizeVertex(new long[] {1}, 1e-10), "bottleneck")
                         // Output
                         .addLayer("outputLayer",
                                         new CenterLossOutputLayer.Builder()
